+++ code/application/game_logic/actor_model.hpp
@@ -35,7 +35,7 @@ private:
   void                  on_start() override;
   void                  on_end() override;
   void                  on_update(const float dt) override;
-  void                  on_event(const uint32_t id, const void *data, const size_t size_of_data) override;
+  void                  on_event(const uint32_t id, const void *data, const uint32_t size_of_data) override;
   
   void                  move_forward(const float fwd);
   void                  move_left(const float left);
