+++ code/application/game_logic/local_player_controller.hpp
@@ -17,7 +17,7 @@ public:
   
   void                  on_start() override;
   void                  on_update(const float dt) override;
-  void                  on_event(const uint32_t id, const void *data, const size_t size_of_data) override;
+  void                  on_event(const uint32_t id, const void *data, const uint32_t size_of_data) override;
 
 
 private:
