+++ code/data/world_data/audio_pool.cpp
@@ -8,7 +8,7 @@ namespace World_data {
 void
 audio_pool_init(Audio_pool *pool)
 {
-  for(size_t i = 0; i < pool->size; ++i)
+  for(uint32_t i = 0; i < pool->size; ++i)
   {
     pool->id[i] = i;
   }
